modify:group_services_function
