+++ src/services/party.ts
@@ -1,11 +1,10 @@
 import { useEffect, useState, useCallback } from 'react'
 import { useNavigation } from 'react-navigation-hooks'
 import firebase, { functions } from '../repositories/firebase'
-import { Party, buildParty, User } from '../entities'
+import { Party, buildParty } from '../entities'
 import { getUser } from '../repositories/user'
-import { showEntryPartyApplySunccessMessage, showEntryPartyApplyFailurMessage } from '../services/flashCard'
 import { useAuthState, useUIActions, useRoomActions } from '../store/hooks'
-import { getParty, setParty, setPartyMembers, setPartyOrganizer } from '../repositories/party'
+import { getParty, setParty } from '../repositories/party'
 
 const db = firebase.firestore()
 const partiesRef = db.collection('parties')
@@ -94,20 +93,5 @@ export const entryDemoParty = async (partyID: string) => {
   } catch (e) {
     console.warn(e)
     return { success: false, error: e }
-  }
-}
-
-export const entryPartyMembers = async (organizer: User, members: User[], partyID: string) => {
-  const batch = db.batch()
-  if (!members || !partyID) return
-  await setPartyOrganizer(organizer, partyID, batch)
-  await setPartyMembers(organizer, members, partyID, batch)
-
-  try {
-    await batch.commit()
-    showEntryPartyApplySunccessMessage()
-  } catch (e) {
-    showEntryPartyApplyFailurMessage()
-    console.warn(e)
   }
 }
